consolidate SP IDs
