+++ src/ccpa/index.ts
@@ -2,9 +2,8 @@ import {
 	init as initSourcepoint,
 	willShowPrivacyMessage as sourcepointWillShowPrivacyMessage,
 } from './sourcepoint';
+import { VENDOR_LIST_ID_CCPA } from '../lib/sourcepointConfig';
 import { mark } from '../lib/mark';
-
-const VENDOR_LIST_ID = '5eba7ef78c167c47ca8b433d';
 
 const init = () => {
 	mark('cmp-ccpa-init');
@@ -15,7 +14,7 @@ const willShowPrivacyMessage = () => sourcepointWillShowPrivacyMessage;
 
 function showPrivacyManager() {
 	// eslint-disable-next-line no-underscore-dangle
-	window._sp_ccpa?.loadPrivacyManagerModal?.(null, VENDOR_LIST_ID);
+	window._sp_ccpa?.loadPrivacyManagerModal?.(null, VENDOR_LIST_ID_CCPA);
 }
 
 export const CCPA: SourcepointImplementation = {
