+++ src/ccpa/sourcepoint.ts
@@ -3,7 +3,7 @@
 import { stub } from './stub';
 import { mark } from '../lib/mark';
 import { isGuardianDomain } from '../lib/domain';
-import { ACCOUNT_ID } from '../lib/accountId';
+import { ACCOUNT_ID } from '../lib/sourcepointConfig';
 import { invokeCallbacks } from '../onConsentChange';
 
 let resolveWillShowPrivacyMessage: Function | undefined;
