--- src/lib/accountId.ts
@@ -1,1 +1,0 @@
-export const ACCOUNT_ID = 1257;
