+++ src/lib/sourcepointConfig.ts
@@ -1,0 +1,3 @@
+export const ACCOUNT_ID = 1257;
+export const VENDOR_LIST_ID_CCPA = '5eba7ef78c167c47ca8b433d';
+export const CMP_ID_TCFV2 = 106842;
