+++ src/tcfv2/index.ts
@@ -2,9 +2,8 @@ import {
 	init as initSourcepoint,
 	willShowPrivacyMessage as sourcepointWillShowPrivacyMessage,
 } from './sourcepoint';
+import { CMP_ID_TCFV2 } from '../lib/sourcepointConfig';
 import { mark } from '../lib/mark';
-
-const SOURCEPOINT_CMP_ID = 106842;
 
 const init = () => {
 	mark('cmp-tcfv2-init');
@@ -15,7 +14,7 @@ const willShowPrivacyMessage = () => sourcepointWillShowPrivacyMessage;
 
 function showPrivacyManager() {
 	// eslint-disable-next-line no-underscore-dangle
-	window._sp_?.loadPrivacyManagerModal?.(SOURCEPOINT_CMP_ID);
+	window._sp_?.loadPrivacyManagerModal?.(CMP_ID_TCFV2);
 }
 
 export const TCFv2: SourcepointImplementation = {
