+++ src/tcfv2/sourcepoint.ts
@@ -2,7 +2,7 @@
 
 import { stub } from './stub';
 import { mark } from '../lib/mark';
-import { ACCOUNT_ID } from '../lib/accountId';
+import { ACCOUNT_ID } from '../lib/sourcepointConfig';
 import { isGuardianDomain } from '../lib/domain';
 import { invokeCallbacks } from '../onConsentChange';
